Move to qrScanActivity
